Daily backup.
This is the final release of the free demo version.
No additional features will be added to this project.

Development will continue under a new name and repository as a professional software product, available with a commercial license and a 3-year software warranty.
If you’re interested, feel free to contact me at attila@tux-net.hu.